feat: 蓝盾对接权限中心RBAC #7794

+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/permission/RbacPipelinePermissionService.kt
@@ -124,7 +124,7 @@ class RbacPipelinePermissionService @Autowired constructor(
         if (iamInstanceList.contains("*")) {
             pipelineInfoDao.searchByProject(dslContext, projectId)?.map { pipelineIds.add(it.pipelineId) }
         } else {
-            val ids = iamInstanceList.map { it.toInt() }
+            val ids = iamInstanceList.map { it.toLong() }
             pipelineInfoDao.getPipelineByAutoId(dslContext, ids, projectId).map { pipelineIds.add(it.pipelineId) }
         }
         return pipelineIds
